feat(environment): remove delete and update e2e
