--- apps/api/src/app/environments/usecases/delete-environment/delete-environment.command.ts
@@ -1,8 +1,0 @@
-import { IsDefined, IsMongoId } from 'class-validator';
-import { OrganizationCommand } from '../../../shared/commands/organization.command';
-
-export class DeleteEnvironmentCommand extends OrganizationCommand {
-  @IsDefined()
-  @IsMongoId()
-  _id: string;
-}
