--- apps/api/src/app/environments/usecases/delete-environment/delete-environment.usecase.ts
@@ -1,15 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { EnvironmentEntity, EnvironmentRepository } from '@novu/dal';
-
-import { DeleteEnvironmentCommand } from './delete-environment.command';
-
-@Injectable()
-export class DeleteEnvironment {
-  constructor(private environmentRepository: EnvironmentRepository) {}
-
-  async execute(command: DeleteEnvironmentCommand) {
-    return await this.environmentRepository.delete({
-      _id: command._id,
-    });
-  }
-}
