+++ apps/api/src/app/environments/usecases/update-environment/update-environment.command.ts
@@ -4,7 +4,7 @@ import { OrganizationCommand } from '../../../shared/commands/organization.comma
 export class UpdateEnvironmentCommand extends OrganizationCommand {
   @IsDefined()
   @IsMongoId()
-  _id: string;
+  environmentId: string;
 
   @IsOptional()
   @IsString()
